WIP Create Faberbabel namespace
