+++ Faberbabel/Classes/String+Translation.swift
@@ -16,8 +16,9 @@ extension String {
     public func fb_translate(to lang: String) -> String {
         var localized = fb_localize(to: lang)
         if localized == self {
-            let event = Event(type: .missingKey, key: self)
-            RemoteEventLogger.shared?.log(event)
+            // TODO: (Pierre Felgines) 08/09/2020 Log event
+//            let event = Event(type: .missingKey, key: self)
+//            RemoteEventLogger.shared?.log(event)
             if lang != "en" {
                 localized = fb_localize(to: "en")
             }
@@ -26,12 +27,14 @@ extension String {
     }
 
     func fb_localize(to lang: String) -> String {
-        let url = Bundle.updatedLocalizablesBundle?.localizableDirectoryUrl?.appendingPathComponent("\(lang).lproj")
-        guard
-            let directoryUrl = url,
-            let bundle = Bundle(path: directoryUrl.path) else {
-                return NSLocalizedString(self, comment: "")
-        }
-        return bundle.localizedString(forKey: self, value: nil, table: nil)
+        // TODO: (Pierre Felgines) 08/09/2020 Return correct string
+        return self
+//        let url = Bundle.updatedLocalizablesBundle?.localizableDirectoryUrl?.appendingPathComponent("\(lang).lproj")
+//        guard
+//            let directoryUrl = url,
+//            let bundle = Bundle(path: directoryUrl.path) else {
+//                return NSLocalizedString(self, comment: "")
+//        }
+//        return bundle.localizedString(forKey: self, value: nil, table: nil)
     }
 }
